rename organization, documentation fixes, update deploy script
