+++ LICENSE
@@ -1,6 +1,6 @@
     GeyserUpdater
 
-    Copyright (C) 2021 Jens
+    Copyright (C) 2021 ProjectG
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
