+++ pom.xml
@@ -4,7 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.alysaa</groupId>
+    <groupId>com.projectg</groupId>
+    <name>GeyserUpdater</name>
     <artifactId>GeyserUpdater</artifactId>
     <version>1.5.0</version>
