+++ src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
-name: GeyserUpdater
-main: com.alysaa.geyserupdater.spigot.SpigotUpdater
+name: ${project.name}
+main: com.projectg.geyserupdater.spigot.SpigotUpdater
 version: ${project.version}
 depend: [Geyser-Spigot]
 description: Automatically or manually downloads new builds of Geyser and applies them on server restart.
